refactor: extract sanitizer into standalone package
