+++ package.json
@@ -11,9 +11,9 @@
     "plugins/*"
   ],
   "scripts": {
-    "build": "pnpm --filter @loamlog/core run build && pnpm --filter @loamlog/archive run build && pnpm --filter @loamlog/distiller-sdk run build && pnpm --filter @loamlog/distill run build && pnpm --filter @loamlog/distiller-pitfall-card run build && pnpm --filter @loamlog/distiller-issue-draft run build && pnpm --filter @loamlog/sink-file run build && pnpm --filter @loamlog/provider-opencode run build && pnpm --filter @loamlog/provider-claude-code run build && pnpm --filter @loamlog/cli run build && pnpm --filter @loamlog/plugin-opencode run build",
+    "build": "pnpm --filter @loamlog/core run build && pnpm --filter @loamlog/sanitizer run build && pnpm --filter @loamlog/archive run build && pnpm --filter @loamlog/distiller-sdk run build && pnpm --filter @loamlog/distill run build && pnpm --filter @loamlog/distiller-pitfall-card run build && pnpm --filter @loamlog/distiller-issue-draft run build && pnpm --filter @loamlog/sink-file run build && pnpm --filter @loamlog/provider-opencode run build && pnpm --filter @loamlog/provider-claude-code run build && pnpm --filter @loamlog/cli run build && pnpm --filter @loamlog/plugin-opencode run build",
     "test": "find packages plugins -name '*.test.ts' -not -path '*/node_modules/*' | xargs node --import tsx --test",
-    "typecheck": "pnpm --filter @loamlog/core run build && pnpm --filter @loamlog/archive run build && pnpm --filter @loamlog/distiller-sdk run build && pnpm --filter @loamlog/distill run build && pnpm --filter @loamlog/provider-opencode run build && pnpm --filter @loamlog/provider-claude-code run build && pnpm --filter @loamlog/core run typecheck && pnpm --filter @loamlog/archive run typecheck && pnpm --filter @loamlog/distiller-sdk run typecheck && pnpm --filter @loamlog/distill run typecheck && pnpm --filter @loamlog/distiller-pitfall-card run typecheck && pnpm --filter @loamlog/distiller-issue-draft run typecheck && pnpm --filter @loamlog/sink-file run typecheck && pnpm --filter @loamlog/provider-opencode run typecheck && pnpm --filter @loamlog/provider-claude-code run typecheck && pnpm --filter @loamlog/cli run typecheck && pnpm --filter @loamlog/plugin-opencode run typecheck"
+    "typecheck": "pnpm --filter @loamlog/core run build && pnpm --filter @loamlog/sanitizer run build && pnpm --filter @loamlog/archive run build && pnpm --filter @loamlog/distiller-sdk run build && pnpm --filter @loamlog/distill run build && pnpm --filter @loamlog/provider-opencode run build && pnpm --filter @loamlog/provider-claude-code run build && pnpm --filter @loamlog/core run typecheck && pnpm --filter @loamlog/sanitizer run typecheck && pnpm --filter @loamlog/archive run typecheck && pnpm --filter @loamlog/distiller-sdk run typecheck && pnpm --filter @loamlog/distill run typecheck && pnpm --filter @loamlog/distiller-pitfall-card run typecheck && pnpm --filter @loamlog/distiller-issue-draft run typecheck && pnpm --filter @loamlog/sink-file run typecheck && pnpm --filter @loamlog/provider-opencode run typecheck && pnpm --filter @loamlog/provider-claude-code run typecheck && pnpm --filter @loamlog/cli run typecheck && pnpm --filter @loamlog/plugin-opencode run typecheck"
   },
   "devDependencies": {
     "@types/node": "^22.15.30",
