+++ packages/cli/src/daemon.ts
@@ -1,17 +1,8 @@
 import { createServer, type IncomingMessage, type Server, type ServerResponse } from "node:http";
 import type { AddressInfo } from "node:net";
 import { writeSessionSnapshot } from "@loamlog/archive";
-import {
-  CAPTURE_PATH,
-  DEFAULT_DAEMON_HOST,
-  DEFAULT_DAEMON_PORT,
-  applySnapshotRedaction,
-  buildSessionSnapshot,
-  parseRedactIgnore,
-  type CaptureRequest,
-  type SessionProvider,
-  isCaptureRequest,
-} from "@loamlog/core";
+import { CAPTURE_PATH, DEFAULT_DAEMON_HOST, DEFAULT_DAEMON_PORT, buildSessionSnapshot, type CaptureRequest, type SessionProvider, isCaptureRequest } from "@loamlog/core";
+import { applySnapshotRedaction, parseRedactIgnore } from "@loamlog/sanitizer";
 import { createOpencodeSessionProvider } from "@loamlog/provider-opencode";
 
 export interface StartDaemonOptions {
